simple-survey/pyproject.toml
[project]
name = "survey-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Krzysztof Kokot",email = "krzysiukokot@interia.pl"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.25,<0.4.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "openai (>=1.86.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langgraph-supervisor (>=0.0.27,<0.0.28)",
    "langgraph-cli[inmem] (>=0.3.3,<0.4.0)",
    "langchain-openai (>=0.3.25,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
